Making Matrix equal to Leia
